Refactor: Move Github to logic
