+++ core/src/com/unciv/ui/screens/modmanager/ModInfoAndActionPane.kt
@@ -14,7 +14,7 @@ import com.unciv.ui.components.extensions.toCheckBox
 import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.components.extensions.toTextButton
 import com.unciv.ui.components.input.onClick
-import com.unciv.ui.screens.pickerscreens.Github
+import com.unciv.logic.github.Github
 import com.unciv.utils.Concurrency
 import kotlin.math.max
