+++ core/src/com/unciv/ui/screens/modmanager/ModUIData.kt
@@ -1,12 +1,9 @@
 package com.unciv.ui.screens.modmanager
 
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton
 import com.unciv.models.metadata.ModCategories
 import com.unciv.models.ruleset.Ruleset
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.extensions.toLabel
-import com.unciv.ui.components.extensions.toTextButton
-import com.unciv.ui.screens.pickerscreens.Github
+import com.unciv.logic.github.Github
 
 /** Helper class holds combined mod info for ModManagementScreen, used for both installed and online lists.
  *
